Factor PATCH /requests into domain layer
