+++ requests_queue/domain/utils.py
@@ -1,0 +1,16 @@
+def deep_merge(source, destination: dict):
+    """
+    Merge source dict into destination dict recursively.
+    """
+
+    def _deep_merge(a, b):
+        for key, value in a.items():
+            if isinstance(value, dict):
+                node = b.setdefault(key, {})
+                _deep_merge(value, node)
+            else:
+                b[key] = value
+
+        return b
+
+    return _deep_merge(source, dict(destination))
